Move loading of ldap settings to LdapComponent
